Entity List / Sheet draft
A visual only implementation of the EntityList and EntitySheet components, combined in an entity browsing page.

This is the starting point to implement the Bureaucrat app. It provides the base configuration for React, how to bundle it using webpack, how to code it using typescript and TSX, how to style it using CSS Modules.

The User niche was chosen because it is so common on so many system. Probably the User and Permissions will be chosen as the first introspected model for the same reason and because the Bureaucrat app itself will need something similar to manage its own permissions.