Introducing Ant Design
The idea of implementing our own design and CSS would cost too much. Let's use something ready to get things done quickly. Maybe in the future a custom design may fit.

The `antd` dependency was added and applied to the whole project, along with a major design change. Mostly style, the general layout of the components remain the same.

Most of the previous `layout` module was deleted.

Also:
- Code format: using single quotes wherever possible.
- Bug fix: domainAuth returns number instead of string.