Support for API Data Domain Instrospection
Mostly changes to allow the new EntityType schema resulted from the efforts of data domain introspection. For instance, an entity may have more than one field acting together to uniquely identify an entity.

A less related change was the support for title patterns (instead of a list of fields to be joined is a space separator). This is more versatile and not a lot harder to implement.

Also, when the user switches EntityTypes, the current Entity is deselected (otherwise there would be a mismatch between the left and right panels.