Editable entity sheet forms
Not yet persisting the new data into the API. That should be the next step.

Handling entity data in React was getting cumbersome it being a class, so the type Entity was simplified into a plain JS object. All of its previous method have been moved to EntityMeta: validation, key generation and title formatting.

Also:
- Simplified some variables, properties and parameters names.
- Fixed some leftovers example domain reference: variable name and user icon.