Delete Entity
* Delete confirmation box

Introduces Ant Modals to deliver confirmation boxes.

In order to decouple the project from AntD as much as possible, AntD's modal methods were not used (https://ant.design/components/modal#modalmethod). Instead, a similar functionality was implemented in `notifications.tsx` and can be accessed via:

```
  // Reack hook:
  const dialogs = useDialogs()

  // Inline dialog confirmation:
  const confirmed = await dialogs.confirm(…)
```

* Forwarding deletion to API

After the deletion completes, the selected entity is set to `null` and the entities are reloaded.